requirements.txt
google-genai
gradio
dotenv